add cookie banner and manage cookie footer
