+++ website/configs/plugins.js
@@ -15,11 +15,7 @@ const setupPlugins = function ({ root, docPath }) {
                 routeBasePath: 'demos',
                 breadcrumbs: false,
             },
-        ], [
-            path.resolve(root, 'src/plugins/wcp-consent-plugin'),
-            {},
-        ]
-
+        ],
     ]
 
     // only add gtag in production
@@ -27,7 +23,7 @@ const setupPlugins = function ({ root, docPath }) {
     const gtmTrackingID = process.env.gtmTrackingID
     if (trackingID && trackingID !== "" && gtmTrackingID && gtmTrackingID !== "") {
         const gtagPlugin = [
-            path.resolve(root, 'src/plugins/visitor-behavior-tracking-plugin'),
+            path.resolve(root, 'src/plugins/visitor-tracking-plugin'),
             {
                 trackingID,
                 gtmTrackingID,
@@ -35,7 +31,7 @@ const setupPlugins = function ({ root, docPath }) {
             },
         ]
         plugins.push(gtagPlugin)
-        console.log(`Add gtag plugin Successfully. tracking ID: ${trackingID}, GTM tracking ID: ${gtmTrackingID}`)
+        console.log(`Add visitor-tracking-plugin Successfully. tracking ID: ${trackingID}, GTM tracking ID: ${gtmTrackingID}`)
     } else {
         console.warn('Failed to get "trackingID" for gtag! Should be null in development, but should be non empty string in Production.')
     }
