--- website/src/plugins/wcp-consent-plugin/index.js
@@ -1,23 +1,0 @@
-module.exports = function pluginWcpConsent() {
-    return {
-        name: 'wcp-consent-plugin',
-
-        getClientModules() {
-            return ['./wcpConsent'];
-        },
-
-        injectHtmlTags() {
-            return {
-                headTags: [
-                    {
-                        tagName: 'script',
-                        attributes: {
-                            async: true,
-                            src: `https://wcpstatic.microsoft.com/mscc/lib/v2/wcp-consent.js`,
-                        },
-                    },
-                ]
-            }
-        }
-    }
-}
